move steps to dedicated folder
