--- src/routes/apartments/[id=apartment_id_matcher]/components/stepper/second-step.ts
@@ -1,6 +1,0 @@
-import { z } from "zod"
-import { NewGuestSchema } from "~/models"
-
-export const SecondStepSchema = z.object({
-    guest: NewGuestSchema
-})
